added layouts, routes, pages
