+++ src/pages/Calendar.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Calendar() {
+  return <div>Calendar</div>;
+}
