+++ src/pages/Company.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Company() {
+  return <div>Company</div>;
+}
