+++ src/pages/Documents.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Documents() {
+  return <div>Documents</div>;
+}
