+++ src/pages/Profile.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Profile() {
+  return <div>Profile</div>;
+}
