+++ src/pages/Reports.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Reports() {
+  return <div>Reports</div>;
+}
