+++ src/utils/classNames.js
@@ -1,0 +1,3 @@
+export default function classNames(...classes) {
+  return classes.filter(Boolean).join(" ");
+}
